+++ package.json
@@ -12,6 +12,7 @@
     "dotenv": "^16.4.7",
     "ejs": "^3.1.10",
     "express": "^4.21.2",
+    "multer": "^1.4.5-lts.1",
     "nodemon": "^3.1.9"
   }
 }
