+++ views/upload.ejs
@@ -9,7 +9,7 @@
 <body>
     <%- include('./navbar.ejs') %>
     <div class="upload-container">
-        <form action="/upload" method="POST" id="upload-form" >
+        <form action="/upload" method="POST" enctype="multipart/form-data" id="upload-form" >
             <input type="file" name="file" required>
             <button type="submit" id="upload-btn">Upload File</button>
         </form>
